requirements.txt

# Django NonRel
https://github.com/django-nonrel/django/archive/nonrel-1.6.zip

# Djang NonRel Dependencies
djangotoolbox
django-mongodb-engine
permission-backend-nonrel

# Django Rest Framework
djangorestframework
django-filter
django-rest-framework-mongoengine
django-rest-swagger

# Django Dependencies
dj-static
markdown

# Web Server
gunicorn

# Unused Git Repositories
#-e git+https://github.com/django-nonrel/django#egg=django
#-e git+https://github.com/django-nonrel/djangotoolbox#egg=djangotoolbox
#-e git+https://github.com/django-nonrel/mongodb-engine#egg=mongodb-engine
#-e git+https://github.com/django-nonrel/django-permission-backend-nonrel#egg=django-permission-backend-nonrel
